requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
zhipuai>=2.1.0
python-docx>=1.1.0
jinja2>=3.1.0
python-multipart>=0.0.9
